+++ examples/todomvc-react/src/components/TodoItem.spec.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import TestUtils from 'react-addons-test-utils'
+import TestUtils from 'react-dom/test-utils'
 import TodoItem from './TodoItem'
 import TodoTextInput from './TodoTextInput'
