+++ examples/todomvc-react/src/constants/TodoFilters.js
@@ -1,3 +1,9 @@
 export const SHOW_ALL = 'show_all'
 export const SHOW_COMPLETED = 'show_completed'
 export const SHOW_ACTIVE = 'show_active'
+
+export const TODO_FILTERS = {
+  [SHOW_ALL]: () => true,
+  [SHOW_ACTIVE]: ({ completed }) => !completed,
+  [SHOW_COMPLETED]: ({ completed }) => completed,
+}
